requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
dnspython==2.0.0
dominate==2.5.2
ecdsa==0.14.1
elasticsearch==7.9.1
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
guess-language-spirit==0.5.3
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
oauthlib==3.1.0
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
python-oauth2==1.1.1
pytz==2020.1
rauth==0.7.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.18
uritemplate==3.0.1
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
